Create update averages trigger
you have to paste this code into a special spot on workbench. Go to the grades table, right click choose 'alter table'
then at the bottom click the 'triggers' tab, go to 'BEFORE INSERT' click the plus sign on the right. now insert this script into the box
between the BEGIN and END statements and run it (click lightning bolt)